Move devices list from account to user devices
