+++ app/src/main/java/android/iot/AccountActivity.kt
@@ -4,20 +4,9 @@ import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import android.view.View
-import android.widget.ArrayAdapter
 import android.widget.ImageButton
-import android.widget.ListView
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.lifecycleScope
-import io.ktor.client.HttpClient
-import io.ktor.client.engine.cio.CIO
-import io.ktor.client.request.request
-import io.ktor.client.statement.bodyAsText
-import kotlinx.coroutines.launch
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.jsonArray
-import kotlinx.serialization.json.jsonObject
 
 class AccountActivity : AppCompatActivity() {
     companion object {
@@ -26,20 +15,6 @@ class AccountActivity : AppCompatActivity() {
         const val LOGGED_IN = "loggedIn"
     }
 
-    suspend fun sendListDevicesRequest(username: String) : Map<String, String> {
-        val apiUrl = "https://vye4bu6645.execute-api.eu-north-1.amazonaws.com/default"
-        val devicesUrl = "$apiUrl/devices"
-
-        val response = HttpClient(CIO).request(devicesUrl) {
-            method = io.ktor.http.HttpMethod.Get
-            headers.append("Content-Type", "application/json")
-            url { parameters.append("username", username) }
-        }
-
-        val responseMap = Json.parseToJsonElement(response.bodyAsText()).jsonObject.toMap()
-
-        return responseMap.mapValues { it.value.toString() }
-    }
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
@@ -48,8 +23,7 @@ class AccountActivity : AppCompatActivity() {
 
         if (!userLoggedIn) {
             val intentLogin = Intent(
-                this@AccountActivity,
-                LoginActivity::class.java
+                this@AccountActivity, LoginActivity::class.java
             )
             this@AccountActivity.startActivity(intentLogin)
         } else {
@@ -63,8 +37,7 @@ class AccountActivity : AppCompatActivity() {
             val backButton = findViewById<View>(R.id.backButton) as ImageButton
             backButton.setOnClickListener {
                 val intentMain = Intent(
-                    this@AccountActivity,
-                    MainActivity::class.java
+                    this@AccountActivity, MainActivity::class.java
                 )
                 this@AccountActivity.startActivity(intentMain)
                 Log.i("Content ", " Main layout ")
@@ -73,56 +46,15 @@ class AccountActivity : AppCompatActivity() {
             val logoutButton = findViewById<View>(R.id.btnLogout)
             logoutButton.setOnClickListener {
                 val intentMain = Intent(
-                    this@AccountActivity,
-                    MainActivity::class.java
+                    this@AccountActivity, MainActivity::class.java
                 )
                 val editor = sharedPreferences.edit()
                 editor.putBoolean(LOGGED_IN, false)
+                editor.putString(USERNAME, "")
                 editor.apply()
                 this@AccountActivity.startActivity(intentMain)
                 Log.i("Content ", "Login layout")
             }
-
-            lifecycleScope.launch {
-                var response = sendListDevicesRequest(username!!)
-                var success = response["success"] == "true"
-
-                if (!success) {
-                    Log.i("Content ", "Failed to list devices")
-                    return@launch
-                } else {
-                    Log.i("Content ", "Successfully listed devices")
-                    var data = response["data"]!!
-                    val deviceNameToId = getDeviceNameToDeviceIdMap(data)
-
-                    var devicesListView = findViewById<ListView>(R.id.lvDevices)
-                    val arrayAdapter: ArrayAdapter<*>
-                    arrayAdapter = ArrayAdapter(this@AccountActivity, R.layout.basic_list_element, deviceNameToId.keys.toList())
-                    devicesListView.adapter = arrayAdapter
-
-                    devicesListView.setOnItemClickListener { parent, view, position, id ->
-                        val element = arrayAdapter.getItem(position)
-                        val intent = Intent(this@AccountActivity, DeviceReadingsActivity::class.java)
-                        intent.putExtra("device_id", deviceNameToId[element])
-                        this@AccountActivity.startActivity(intent)
-                        Log.i("Content ", "element " + element + deviceNameToId[element])
-                    }
-
-                }
-
-            }
         }
-    }
-
-    private fun getDeviceNameToDeviceIdMap(data: String): Map<String, String> {
-        var devices =
-            Json.parseToJsonElement(data).jsonArray.map { it.jsonObject.toMap() }
-        var devicesMap = buildMap<String, String> {
-            for ((index, device) in devices.withIndex()) {
-                var deviceId = device["device_id"].toString().replace("\"", "")
-                put("Device " + index.toString(), deviceId)
-            }
-        }
-        return devicesMap
     }
 }
