+++ app/src/main/java/android/iot/UserDevices.kt
@@ -2,15 +2,21 @@ package android.iot
 
 import android.app.AlertDialog
 import android.content.Intent
-import android.iot.bluetooth.BluetoothListDeviceAdapter
-import android.iot.bluetooth.Data
-import android.iot.bluetooth.RecyclerViewClickListener
 import android.iot.devices.Device
 import android.iot.devices.UserListDeviceAdapter
 import android.iot.devices.UserRecyclerViewClickListener
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.ImageButton
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.lifecycleScope
+import io.ktor.client.HttpClient
+import io.ktor.client.engine.cio.CIO
+import io.ktor.client.request.request
+import io.ktor.client.statement.bodyAsText
+import kotlinx.coroutines.launch
+import kotlinx.serialization.json.Json
+import kotlinx.serialization.json.jsonArray
+import kotlinx.serialization.json.jsonObject
 
 class UserDevices : AppCompatActivity() {
 
@@ -38,18 +44,18 @@ class UserDevices : AppCompatActivity() {
         val addButton = findViewById<ImageButton>(R.id.addDevice)
         addButton.setOnClickListener {
             if (getSharedPreferences(SHARED_PREFS, MODE_PRIVATE).getBoolean(LOGGED_IN, false)) {
-                val intentMain = Intent(
+                val intentPairedDeviceList = Intent(
                     this@UserDevices,
                     PairedDeviceListActivity::class.java
                 )
 
-                intentMain.putExtra("devicesAddresses", devices)
-                this@UserDevices.startActivity(intentMain)
+                intentPairedDeviceList.putExtra("devicesAddresses", devices)
+                this@UserDevices.startActivity(intentPairedDeviceList)
             } else {
                 //  User is not logged in
                 val builder: AlertDialog.Builder = AlertDialog.Builder(this)
                 builder
-                    .setMessage("Pleas log in to add a device!")
+                    .setMessage("Please log in to add a device!")
                     .setTitle("Not logged in!")
 
                 val dialog: AlertDialog = builder.create()
@@ -57,41 +63,70 @@ class UserDevices : AppCompatActivity() {
             }
         }
 
-        val data = this.getData()
+        lifecycleScope.launch {
+            val data = this@UserDevices.getData()
 
-        val listener = object: UserRecyclerViewClickListener() {
-            override fun onClick(index: Int) {
-                super.onClick(index)
+            val listener = object: UserRecyclerViewClickListener() {
+                override fun onClick(index: Int) {
+                    super.onClick(index)
 
-                val intentMain = Intent(
-                    this@UserDevices,
-                    DeviceReadingsActivity::class.java
-                )
+                    val intentDeviceReadings = Intent(
+                        this@UserDevices,
+                        DeviceReadingsActivity::class.java
+                    )
 
-                intentMain.putExtra("device_id", data[index].uuid)
+                    intentDeviceReadings.putExtra("device_id", data[index].uuid)
 
-
-                this@UserDevices.startActivity(intentMain)
+                    this@UserDevices.startActivity(intentDeviceReadings)
+                }
             }
+
+            val adapter = UserListDeviceAdapter(data, this@UserDevices, listener)
+
+            val recyclerView = findViewById<androidx.recyclerview.widget.RecyclerView>(R.id.itemsList)
+
+            recyclerView.adapter = adapter
+            recyclerView.layoutManager = androidx.recyclerview.widget.LinearLayoutManager(this@UserDevices)
         }
 
-        val adapter = UserListDeviceAdapter(data, this, listener)
-
-
-        val recyclerView = findViewById<androidx.recyclerview.widget.RecyclerView>(R.id.itemsList)
-
-        recyclerView.adapter = adapter
-        recyclerView.layoutManager = androidx.recyclerview.widget.LinearLayoutManager(this)
     }
 
 
-    private fun getData(): ArrayList<Device> {
+    private suspend fun getData(): ArrayList<Device> {
         val output = ArrayList<Device>()
+        val response = sendListDevicesRequest(getSharedPreferences(SHARED_PREFS, MODE_PRIVATE).getString(USERNAME, "")!!)
+        val devices = getDeviceNameToDeviceIdMap(response["data"]!!)
 
-        output.add(Device("Device 1", "00:00:00:00:00:00", "00:00:00:00:00:00"))
-        output.add(Device("Device 2", "00:00:00:00:00:00", "00:00:00:00:00:00"))
-        output.add(Device("Device 3", "00:00:00:00:00:00", "00:00:00:00:00:00"))
-
+        for (device in devices) {
+            output.add(Device(device.key, device.value, device.value))
+        }
         return output
+    }
+
+    private suspend fun sendListDevicesRequest(username: String) : Map<String, String> {
+        val apiUrl = "https://vye4bu6645.execute-api.eu-north-1.amazonaws.com/default"
+        val devicesUrl = "$apiUrl/devices"
+
+        val response = HttpClient(CIO).request(devicesUrl) {
+            method = io.ktor.http.HttpMethod.Get
+            headers.append("Content-Type", "application/json")
+            url { parameters.append("username", username) }
+        }
+
+        val responseMap = Json.parseToJsonElement(response.bodyAsText()).jsonObject.toMap()
+
+        return responseMap.mapValues { it.value.toString() }
+    }
+
+    private fun getDeviceNameToDeviceIdMap(data: String): Map<String, String> {
+        var devices =
+            Json.parseToJsonElement(data).jsonArray.map { it.jsonObject.toMap() }
+        var devicesMap = buildMap<String, String> {
+            for ((index, device) in devices.withIndex()) {
+                var deviceId = device["device_id"].toString().replace("\"", "")
+                put("Device " + index.toString(), deviceId)
+            }
+        }
+        return devicesMap
     }
 }
