+++ app/src/main/res/layout/activity_account.xml
@@ -52,27 +52,6 @@
             android:layout_marginTop="16dp"
             android:text="Logout" />
 
-        <TextView
-            android:id="@+id/yourDevices"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/btnLogout"
-            android:layout_centerHorizontal="true"
-            android:layout_gravity="center_horizontal"
-            android:layout_marginTop="16dp"
-            android:textStyle="bold"
-            android:textColor="#010203"
-            android:text="Your devices:" />
-
-        <ListView
-            android:id="@+id/lvDevices"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/yourDevices"
-            android:layout_marginTop="10dp"
-            android:textColor="#010203"
-            />
-
     </LinearLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
